Added the Path and Dragon Classes. Added a method to Unit.
The Path class has an array list of Vec2D.Short's. It will be used to
create the points that the units will move through. The Dragon class is
a unit and will be a dragon in the game. The method added to Unit is
setWayPoint. This will be the point that the unit will start to move to.